Finish DisplayCmd

+++ Commons/command/display/command.go
@@ -2,35 +2,41 @@ package display
 
 import (
 	"context"
-	"errors"
 	"net/url"
 	"os/exec"
 
-	"github.com/EliasStar/DashboardUtils/Commons/command"
+	. "github.com/EliasStar/DashboardUtils/Commons/command"
 	"github.com/EliasStar/DashboardUtils/Commons/util/misc"
 )
 
 type DisplayCmd struct {
-	// TODO DisplayCmd url getter and remover
-	URL string
+	Action DisplayAction
+	URL    string
 }
 
 func (d DisplayCmd) IsValid(ctx context.Context) bool {
+	valid := d.Action.IsValid()
 	_, err := url.Parse(d.URL)
-	return err == nil
+	return valid && err == nil
 }
 
-func (d DisplayCmd) Execute(ctx context.Context) command.Result {
+func (d DisplayCmd) Execute(ctx context.Context) Result {
 	cmd, ok := ctx.Value(misc.DisplayContextKey).(*exec.Cmd)
 	if !ok {
-		return command.ErrorRst{errors.New("display not initialized")}
+		return NewErrorRstFromString("display not initialized")
+	}
+
+	if d.Action == ActionGet {
+		return DisplayRst(cmd.Args[1])
 	}
 
 	cmd.Process.Kill()
 	cmd.Process.Release()
 
-	cmd = exec.Command(misc.DashDBrowser, d.URL)
-	cmd.Start()
+	if d.Action == ActionSet {
+		cmd = exec.Command(misc.DashDBrowser, d.URL)
+		cmd.Start()
+	}
 
-	return command.OKRst{}
+	return OKRst{}
 }
